// apps/betty-server/src/ir/domain/enums/ir-input.enum.ts
export enum IrInput {
  ZERO = '0',
  ONE = '1',
  TWO = '2',
  THREE = '3',
  FOUR = '4',
  FIVE = '5',
  SIX = '6',
  SEVEN = '7',
  EIGHT = '8',
  NINE = '9',
  STAR = '*',
  HASH = '#',
  OK = 'OK',
  UP = 'UP',
  DOWN = 'DOWN',
  LEFT = 'LEFT',
  RIGHT = 'RIGHT',
}

export const NEC_COMMAND_TO_INPUT: Record<number, IrInput> = {
  0x19: IrInput.ZERO,
  0x45: IrInput.ONE,
  0x46: IrInput.TWO,
  0x47: IrInput.THREE,
  0x44: IrInput.FOUR,
  0x40: IrInput.FIVE,
  0x43: IrInput.SIX,
  0x07: IrInput.SEVEN,
  0x15: IrInput.EIGHT,
  0x09: IrInput.NINE,
  0x16: IrInput.STAR,
  0x0d: IrInput.HASH,
  0x1c: IrInput.OK,
  0x18: IrInput.UP,
  0x52: IrInput.DOWN,
  0x08: IrInput.LEFT,
  0x5a: IrInput.RIGHT,
};